899 | L10N - Extend the UI language chooser
https://trello.com/c/HaXP1K3L/899-l10n

This changes causes that if a bundle has messages.properties
file in some language, this language will be visible and can
be selected by user on startup page and after log in on
dashboard page.

Because of possibility of having a lot of languages on startup
page the UI language chooser is changed from radio buttons to
dropdown list.

I added several improvements to the user interface
 - Improved space in the dropdown button
 - Equaled space in the radio buttons
 - I made the startup panel more responsive
 - Increased the input field

Change-Id: I0c19c4edba14a703a574c9b2af4d45837b7d8428