String format manipulation feature in tasks
https://trello.com/c/8aDdvys9

Added new string manipulation for string trigger
and data sources fields. Format manipulation
works similarly to String.format method. Field
to which we want to add 'format' manipulation
must contain format of message where each '%s'
will be changed to value which is in 'format'
manipulation. The order of values in
manipulation is very important.

For now the manipulation is available only for
Firefox version of the Tasks module UI and it
supports only string values.

Change-Id: Ic4c32f0d9c69841645e7fa09618d91819a25b01c