plusDays manipulation feature in tasks
https://trello.com/c/lGvdOq08

Added new plusDays manipulation for date trigger and data sources field.
The manipulation works similarly to DateTime.plusDays method. It needs
an argument which presents the number of days which should be added to
date in the field.

For now the manipulation is available only for Firefox version of
the Tasks module UI.

Change-Id: I0bca0d7f7fa4f9ab00357a3d689dda6af1bf7ee9