965 | Changed maven articat to RPM so that it is deployed to nexus.
https://trello.com/c/n67npIIy

Previously, the primary arfifact type was specified as jar. This resulted
in a dummy jar file getting generated as the primary artifact and the
actual rpm was not considered as artifact. So, mvn deploy did not upload
the rpm to nexus. Fixed this issue by changing the primary artifact type
to "rpm" and making the goal to attached-rpm.

Change-Id: I7c3480ed57326d5dfff63dab6f9682669ad1cd6d