BUG: Remove creating motech user from web-security build process.
https://trello.com/c/wF5PQdfd

The motech admin user was not used by the web-security process. But it was
used by the admin tests. So, admin build process creates the user before IT
and deletes the websecurity db after IT. Platform server build process
does the same because the ITs also needs the admin user. As the same user
data was used, the admin user JSON has been moved to parent project base dir.

Change-Id: I84c83ef7e28a840e1b56dd7ccb7e4a0798a481b5