Fixed bug: using failIfDataNotFound flag for data sources in filters as well
"failIfDataNotFound" specifies whether the task should fail when provider
lookup does not return anything. The behaviour is expected when datasource
is used in either filter or action. Currently it's present only when the
datasource is used in an action but not a filter. This fix is to use the
flag when datasource is used in a filter as well.

Some refactoring as part of this card.
- Extracted domain objects out of static helpers in TaskInitializer and
HandlerUtil.
- Moved key expression related functions to KeyInformation, filter checks to
TaskFilterExecutor, type conversion handling to ParameterType.
- Introduced TaskContext that holds data source objects used during
filter/action execution.

https://trello.com/c/NIz6Eurb/

Change-Id: I3dc933e2b5a078301dd5f9f990a52adb9147f69d