WebSecurityBundleIT fails randomly
https://trello.com/c/1ZmdKbBH

Modified the WebSecurityBundleIT test. Before change the test checked
if there are available two services (from config and event modules) and
set ActiveMQ property. There could be a situation when ActiveMQ config
file did not exist and this caused that the test failed.

Currently the test checks if there are available three services from
web security module. Next it tries to add one role with one permission
and then register user with the role. In the end the test checks that
everything were saved correctly.

Change-Id: I663b22265b7274e72625d2fb522d51d57ffa3b89