Squid: Avoid cycle between java packages
https://trello.com/c/A0rrIWb0

Configured the violation to fail the build:
* added enforcer rule (enforcer-maven-plugin)
* used JDepend to find dependency cycles

Deffinition of package may vary. For this issue I've adopted the
Sonar (and JDepend) definition. Note that (for example) for Idea plugin
"package" means package and its subpackages, while for Sonar package
and its subpackages are not the same package.
Example:
packages:
a.v
a.x
b.y
b.z

Sonar (and JDepend) will be searching for cyclic dependencies
between a.v, a.x, b.y and b.z
Idea plugin will be searching for cyclic dependencies between
a and b.

Change-Id: I5bb41a5107a8163a5dc5b9a02b60bbd1a3eea272