Config: Get platform's core settings from File (in file mode)
Checking if config is already stored in the database,
Connecting StartupManager with bootstrap, checking config source from bootstrap
in StartupManager, add platformInitialized field in SettingsRecord.
ConfigFileMonitor listen files changes only in file mode now.
I added also bootstrap.properties and copying it to .motech/config in build time
(It is needed to review this changes and to loading motech until
http://review.motechproject.org/#/c/3237/ don't be merged. I will remove it after that)

https://trello.com/c/GCum6lNP

Change-Id: I6b39556123461078d0363579a3d450594a79df20