Refactored creation of left hand nav. Fixes highlighting bug.
https://trello.com/c/ryxZSBC8/1126

Instead of downloading two separate lists of modules, a single menu
object is now built on the server side. This menu is retrieved on the
UI by an angular service. This not only simplifies the menu processing,
but also fixes the bug in question - wrong links highlighted in the
navigation. Also, while at it, I deleted a duplicate dependency from
the server bundle pom.

Change-Id: I2b1e0567eb22772435bc4c9d1b66745003feb04b