Seuss: Performance testing and improvements
https://trello.com/c/g0xhiwyn

The purpose of this commit is to add a comparision test between two solutions:
- MDS data persistence and old couchDB solution. By default, the test is
disabled, since it runs a long time for bigger amount of instances. To run
performance check locally, you must remove the "ignored" prefix and run the
test via maven. The amount of instances can be easily adjusted via the constant
at the beggining of the test class. As a part of this commit, some
adjustments have been made to improve the performance of the MDS module.

Change-Id: I2ec079bd8d5aff4f38b4b22b89d80fc34c968007