WS: Change how security rules are loaded
https://trello.com/c/V5NNDQAd

Until now, the only way to update the security rules was to force the user
to clean his database or make him add the rules manually themselves.
From this commit on, on the server startup, the default security rules
are checked against what's in the database and update it. To ensure
once deleted rules don't come back, a new property "deleted" has been
added. Rules marked as "deleted" should be treated exactly as if they didn't
exist.

Change-Id: If8da7c253274f92459376bd7db118fee94c8c463