SVH: Added the module trash mode.
https://trello.com/c/iat9Q4sA

From this commit the MDS module will have the trash mode. Basically if
the mode is active, instances that should be removed are moved to module
trash If a user decides to disabled the trash mode, instances will be
removed permanently.

If the user decided to clean the trash after some time internal the
repeating schedule job will be scheduled and in the correct time the
appropriate method will be executed.

Change-Id: I13879915f2e861c4a8b136c841364a6ce348ccd3