MOTECH-793: Support for enum types in MDS (ui only)
Currently when a user will create a field with combobox type, the list
type will be used to create a field in the class definition. This commit
change this behaviour. From now there will be used one of the following
type:

 * String      - if the user supplied option will be set and multiple
                 selections option will not be set,
 * Enum        - if the user supplied and multiple selections options
                 will not be set,
 * Enum list   - if only the multiple selections option will be set,
 * String list - if both options will be set.

The new enum class with name '{entityName}{fieldName}Enum' and in the
same package as entity class will be created when it is necessary and
it will have the same value as those are defined in the field.

Change-Id: Id3d0055f301c887a7b336ddc175508b76040eb98