MOTECH-997: Fix issue with incomplete logging in pax exam tests
This commit removes an outdated dependency on surefire-junit47. The bug
would lie with that dependency, it would hit and index out of bounds
Exception when printing stacktraces. This also fixes an issue with a
test that can fail because of test ordering(adding to a set which is
reused in a different test).

Change-Id: I59bb00ad0c25b907661971b6a2984aef8c821b6d