MOTECH-1031: Added support for locale types in MDS
This commit adds support in the MDS module for java.util.Locale type.
The UI is kept simple for now - user can select a language but will
be displayed as locale code (e.g. 'en' insead of 'English). Adding
fancy UI will be done as a part of separate card.

Change-Id: I9fd15f0766537988535e75ce06486189100d38a7