MOTECH-858 | Update archetype for Seuss
Move some classes around to better match our guidlines. i.e. service api
in services package with the implementation in impl.  Update the
repository archetype so it uses MDS instead of couch and get all the
generated integration tests (except http) passing.  This is an improvement
 over the current archetype where they all fail.  The remaining failure
seems to be web security related.

Change-Id: I5abde4c0643e570452870767e1e0837b37c499ca