MOTECH-1038 BUG: Lookup count always returns 1 for single return lookups
For each lookup we annotate in service interface, MDS generates
a count lookup. However, when the lookup returned a single object,
the count method would always return 1, even if the object didn't
exist. This has been fixed and a proper number should be returned
now. Also a test has been added to verify that.

Change-Id: I7bcb69ac7656f29e75abb0febbd4d92dc78b4e6b