Fix for total record display in MDS grids
All grids in MDS would display wrong information in their footer, which
was always the minimum between the total count and page size, since
list.size() was used for the total record number value(where list was the
data to be displayed). I fix this issue and some explanation to what the
fields in the Records class mean.

Change-Id: I1a2d9ce0d43655f9692dac5ba7e8a59accebfa48