MOTECH-1077: Auto generate values for default fields in entity
Each entity has default set of properties. For some of them the value
should be generate automatically. Manual setting was good for simple
entities but because the MDS module will support relationships this
mechanism had to be upgraded. From now the values are generated by data
nucleus when the instance is created so properties in related entities
also will be generated. To omit problem with updating some of fields all
setters in class definition after set property execute the
updateModificationData method which set modification properties.

Change-Id: Ic9cad32d6da384787d238275095cdf1615aedbde