MOTECH-1058: Pick up JDO annotations from fields in entity classes
Our own implementation of JDOAnnotationReader had to be implemented and
used in DataNucleus. This reader extends the standard one from DataNucleus
and only overrides its behavior of recognizing persistence-capable classes.
For classes annotated with just @Entity, we now return a simple
representation of the @PersistenceCapable annotation. Thanks to this
the annotation reader will commence with reading field-level annotations.
The reader class had to be defined as a DataNucleus extension in plugin.xml.

Change-Id: I755c1e6bb717249dabac363e56658b8c38c0ba62