MOTECH-1086: Added LocalDate type to MDS
This is required for Pillreminder. Support on the backend and UI side.
Moreover I have added some test cases for the Locale type. I have noticed
a few issues with datepickers, by they can be resolved as part of different
tickets, the workarounds on the backend side we have now should suffice.

Change-Id: I741d21b50b2c396e1de2ad03a5a8d8b3997835a7