MOTECH-1081: Allow using custom operators in Lookups
Users can use custom operators in lookups. In order to do
so, they need to add the customOperator parameter to their
@LookupField annotation. Custom operators will be used
instead of ==(equals) checks. Providing methods such
"matches()" is also supported.

Also, InstanceServiceImpl was updated in order to allow
the user to provide emtpy/sets ranges(equivalent to null)
for the users convenience.

Change-Id: Ifecc2dda3473bf73e5ea985386f711af37cae447