Map metadata adjustments
Small adjustments have been applied while generating metadata
for fields of the map type. MDS now stores key and value class
types in the field metadata if the field is of map type. That
allows to choose a strategy when building metadata:
 * Serialize map - for maps other than Map<String, String>
 * Store map in separate table - for Map<String, String>
The second case is most commonly used in Motech. Thanks to
storing map in separate table we get the ability to create
custom queries, what would be impossible for serialized map.

Change-Id: If79589718c6572d60a33de8707e4e5ab893806ba