MOTECH-843: Removed couch dependency from POMs
This is a suqash merge from the branch dev/bye-couch branch.
This removes all(almost) references to couchdb and ektorp.
More specifically this commit:

* Removes antrun plugin definitons used for manipulating couch databases
* Removes couch configuration from the bootstrap config, commons-couchdb
  now uses only properties from its classpath(getting ready to remove it)
* Removes couch configuration from the bootstrap page, removes validation
  code used for couchdb url
* Removes code from the admin module settings management,
  refering to couch(it was unused anyway)
* Removes dependnecies we do not need, such the couchdb quartz jobstore
  and ektorp lucene
* Removes couch dependency from deb/rpm
* Makes couch module no longer used in the Server ITs
* Copies httpcache exclusion same as in modules repository, so we avoid
  a stacktrace during the Server ITs

The commons-couchdb modules is left intact, but no longer uses bootstrap configuration.
It is still is used for MDS performance tests, and I believe it will stay there as a
dependency after it gets move to contrib(test dependency).

Change-Id: I85df2da19067dcea3eac3028a06d9b55cc8a8155