MOTECH-1218: fixed problem with QueryNotUniqueException
If a user creates a new instance with relation field and the user will try
to update instance that is in relation field, a QueryNotUniqueException
will be thrown.

This is caused because for instance from relation field
there is no check whether historical data exist and because that there will
be two records in database with LAST flag.

This commit fixed this issue. When historical data is created for each
relation field is checking if in database there is no historical data and
if there is then for these data LAST flag is unset.

Change-Id: I4a93f40ae539052a45f1c69051c85f98eb83a7f7