MOTECH-1214 | 0.24 release notes
Adds release notes for the 0.24 release. There are still a couple of
TODOs in the docs (links to binaries, etc.). I'll resolve those TODOs
with my next checkin. I've created a placeholder for the TOC page, but
haven't added a link to the rel notes yet - will do that after we release.
Change-Id: Ie1243a087d8eb5342e95439cb0c4638dd99114b5