MOTECH-1164: Fix for random NPEs
The underlying issue lies at Felix. When the Nucleus context is created
it is given ClassLoaders coming from the old bundle version, which leads
to NPE when we try to load resources from those ClasslLoaders. This is a
pretty ugly fix, but it works and seems the best we can do right now. We
catch the NPE and then recreate the ClassLoaderResolver instance(its
classloader field is final).

Change-Id: I77e860578e117c800e273d3c195f24e1a61169ab