MOTECH-1211 Improve performance of history recording
This commit removes the "isLast" flag from the history class,
generated for entities. This flag was used to mark current revision
of the instance, in order to NOT show it on the list of historical
revisions. This approach has been changed and the latest revision
is discovered by the ID now. Thanks to this, we don't have to perform
one additional database write for each object update, what speeds it up,
especially for huge instances with many relations.

Change-Id: I115548d4a0ed2b5ec2d665d421d6713a5b5a87bf