MOTECH-1266: RestOperations annotation support
This commit introduces RestOperations annotation which can be applied to
entity class to set related RestOptions metadata. RestOperations value is
an array of one or more values specified in RestOperation enum: CREATE,
READ, UPDATE, DELETE and ALL. If ALL is specified, all CRUD options are
enabled, regardless of presence of other values.

Change-Id: If9aca0f0b2a5378788e4f8b9d33381a4e0788f3e