MOTECH-1186: In Pax ITs classes are loaded with different ClassLoader
Previously after loading class with entity annotation, we were loading fields
and methods of this class. Unfortunately in Pax ITs there was a problem with
relationships. If field's generic type is instance of related class e.g.
Set<AnotherClass>, then 'AnotherClass' is loaded with wrong ClassLoader.

To be sure that two related classes will be loaded with MDSClassLoader now
we are loading all classes with entity annonation in bundle and after this we
are loading theirs fields and methods.

Change-Id: Ib138508668c85cdee8608ce24be9340b89ee35ab