MOTECH-1284: Distinct MDS behaviour depending on install method.
Generally there is a problem with ManyToMany relation in MDS after MOTECH-1047
commit. It depends on schema creation - precisely on order of columns in JOIN
table. Unfortunately with this implementation it will not work for List
collection. However it can work with Set. More details about ManyToMany in
link below :
http://www.datanucleus.org/products/datanucleus/jdo/orm/many_to_many.html

I have created integration test to show how it should look like. Really
important thing is to override hashCode() and equals() method so that updates
are detected correctly (more info also in link).

Change-Id: I7ffb3e208e7722dfc3712c5d194f83e807ce76e6