MOTECH-1237: TEXT field needed
Added support for TextArea field to MDS. To use it in DDE
String field needs to bo annotated with "type="text". That
field will be handled as String type with TEXT instead of
VARCHAR in DB.

MOTECH-1169 is also done in this commit. When creating EUDE
user can choose TextArea type (which is available only from UI).
It is later handled as normal String type with TextArea setting
set to true. Also option to change that setting on String type
using UI is no longer available (user should use TextArea type instead).

Change-Id: Iacbe91dc278cec45de6f1e5a4dcc0dae263f2bec