MOTECH-1205 Fix occasional duplicate registration of Task channels
I believe the problem here was that multiple threads accessed method
responsible for creating and updating of the tasks channels.
I've put synchronized on the method, what should solve
the issue. Additionaly, checking whether the channel is present
and depending on the result either creating or updating the channel
is now done in a single transaction.

Change-Id: Ic52ad4a432452a33a4b0bb0a9f3a55a16e5a1ff8