Fix task channel registration when installing module from UI
I noticed that when we install a new module via UI, its
channel does not get registered. The reason for that is that when
entities bundle is getting restarted, what unregisters the ChannelService,
the ChannelService did not get registered back on, due to the info
that tracker is already present. This change should fix this and the
channels should get properly registered now.

Change-Id: I4d3aa3e5f24a6e6fa380fc726cb2b9f8ac597b94