MOTECH-1396 Required fields are now validated before storing
Until this moment, when we marked DDE required, it was validated
only on the UI. That meant it was possible to use the data service
for an entity and place null values even in required fields. I've
added JDO metadata for required fields and set it to throw an
exception when an attempt to store null value at required field
is performed. I've also extended MdsDdeBundleIT to test this case.

Change-Id: Ib8d6cc9866fa5228d572ec6f691a5641ef9cc6dc