MOTECH-1369 Adjusted task channels update
A recent change expanding the transaction in the ChannelService
caused problems in some cases of channel updates. It seems an
issue here is that removing objects and then adding the same
objects again in the same transaction doesn't work well with
Datanucleus. To bypass this, we create new instances. I've also
added a simple check before updating the channel - if nothing
has changed, no attempt of update will be performed.

Change-Id: Ibf4252ac8b421e8ed100ed9a55cb117990d28467