MOTECH-1302: Standalone war
The war will now run without any modules in ~/.motech/modules.
Admin and scheduler are the two non platform modules that are
part of the war. Bundles from ~/.motech/bundles can override
the bundles from the war if they can have the same symbolic name
and version. MDS had to be extended to support jndi: urls in
reflections, so a service for this is exposed by the server
code. I also made the scheduler dependency in the email module
completely optional at runtime.

Change-Id: I28e59db0935ee3012799acfa7d918e5f1ad97480