MOTECH-1447 Renamed LocalDate to Date
LocalDate was the only true Date-only type in MDS. The previous
Date type has been renamed to java.util.Date and is still available
in MDS, but for DDE exclusively. java.util.Date type has got a picker
with date and time now on UI. I've also adjusted docs to reflect the
changes.

Change-Id: Ie7f87d412ae98ce4bf3411b881cb1bcc16cd0e42