MOTECH-756 Integrating MOTECH with CommCare topic
Added a documentation on integrating MOTECH with CommCare.
The document explains step by step, how to connect to
the CommCareHQ account and explains the meaning of every
option in the Settings tab. A way to veirfy the settings on
the CommCareHQ account is also described. Moreover, a note
on the fired events has been added as well as a note
on the ability to use Commcare data in task triggers
and data sources.

Change-Id: I65a6cf111c74391e5403417dbe2636af25e3e46c