MOTECH-1476: Added comparator for Dynamic URLs
Rules in UI was in different order than in filter. Also after restarting the server,
the filter order was differed than before. Now the rules are displayed in the
same order as in the filter.

Also changed the sort order:
If priorities are equal then first will be checked rule with origin different than
"SYSTEM_PLATFORM". If both rules have "SYSTEM_PLATFORM" origin then the order is
determined by the URL pattern length(longer will be checked first). If the lenghts
of patterns are equals then is used compareTo method.

Change-Id: Id7b04bb7069c6d79a20668b0f3f0b3dbee5b4234