MOTECH-1359: InnoDB - DataNucleus - Specified key was too long
Changes MySQL file format to Barracuda and turns on large prefixes.
Thanks to that it is possible to use longer keys as long as table
row_format is set to dynamic. Also adds custom MySQL data adapter to
Datanucleus that changes row_format to dynamic to prevent key length
error.

Change-Id: I3cd0a849eba80f40bb6984d2c240fa3ae640074f