MOTECH-1535: Delete always returns 200
We now check if the item we are deleting is null. Moreover I have cleaned
up the code handling deletes. I removed methods which would do a retrieve
then delete in the repository. We still need to retrieve first when
deleting a single instance, since we are also handling trash for it.

Change-Id: Ia54323db5affe6b932f4b4d30171c0abc0677e85