MOTECH-1188: Fixed test failures on Windows
Directories with spaces were the cause of the issues.
Reflections double encodes directory urls in Windows,
which seems a bug with the version we use. I added a
separate handler for this buggy case. Moreover I fixed
test failures in the config modules, caused by not properly
converting these urls to files in these tests.

Change-Id: I9d7c38861f241d00c74504e2f219f21fff5fdcfa