MOTECH-1542: Added entity class to MOTECH CRUD events
I added entity class separately to the payload of the event.
The name of the entity now has an "Entity Name" display name.
I also removed some redundant copy-paste code from the velocity
template for the task-channel and changed the order of parameters
in subject building methods, to make them more consistent.

Change-Id: I939810757d56b026b3fb72dde8a2269f77fff461