MOTECH-1523 Wrap annotation processing into single transaction at startup
The annotation processors do not insert any data to MDS anymore. Instead, they
return a representation of what they managed to discover, by the getProcessingResult()
method. The results coming form the bundles are collected, and at the end everything
is inserted into MDS at once. The rest fields and lookup fields are now discovered based
on the name, not ID on the dto level.

Change-Id: I23aab83b408e0c5eb26446b1b4fc2426b6a8e7e2