MOTECH-968: Integrated Swagger UI into MOTECH
Swagger documentation spec in json format is generated
along with the MDS entities bundle. This documentation is
stored as text in the database. The module mds-web exposes
and endpoint which allows accessing this spec. A mechanism
for registering REST documentation spec endpoints was added
to the server-bundle module(this required some changes to module
loading). Swagger UI itself was added to the server-bundle module
as separate menu on the top, where modules are shown as sub-menus.

Swagger-ui is generated based on the spec retrieved from the
registered endpoint and allows hot testing of the api. The current
MDS spec is generated for all entities that expose REST access.

Documentation for the swagger-ui mechanism, lookup support in the
generated spec and a test for the endpoint in mds-web are out of
scope for now(we have more tickets in this epic).

Change-Id: I16df7c97643bd425e008c0491ceb32ca2f7b2cb2