MOTECH-1585: Fix task dropdown badges with special chars
Tasks would do a match using the display name, which
was bad. I placed the event key as a parameter of the
draggable span, which should solve the issues some
people have run when using the Commcare channel.
I also fixed the message function for rebuilding
the badges - it should now no longer use square
brackets.

Change-Id: I0c8d465ee5283cd8b2eb307e41bd6cadd13932fc