MOTECH-1575: Drafts don't get outdated if new values were added to a combobox.
This commit makes draft outdated when new values are added to a combobox.
When we update entity, automatically version of entity is incremented and
draft has to be updated to newer version. Updating draft works similar to
abandon changes.

Also I have to correct some other bugs.
First of all method updateDraft didn't work correctly. Mainly because of null
value of entity in field and lookup objects.
Secondly lookups were retrieved using getById, which wasn't correct because
the same lookup has different ID in Entity and EntityDraft.
Thirdly when we want to update draft in entity with lookups, every time
LookupNameIsRepeatedException was thrown. Generally there is no way to add
lookup with the same name. On the UI javascript covers this problem and when
we want to add two lookups with the same name using annotation, always the
second one overwrites the first one.
Last but not least, there was a problem with parsing empty default value in
combobox. Function returned empty array instead of empty string, which caused
compiler error in javassist.

Change-Id: Ic60e337f349a43a1cc64e90ebb1fcfab2397ad34