MOTECH-1546: Added lookup support to Swagger + few fixes
Endpoints for lookups are now generated in the
Swagger documentation.

I have also changed the way generation is done.
Now it is done per request in order to support
restricting content per request in the future.

Moreover I fixed an issue with MDS overriding
lookup REST settings for DDEs on restart.

Access to the REST section was untied from UI
permissions.

Lastly, I added a 403 response to all Swagger
endpoints.

Change-Id: If34eda4789ddcfb1e56f091d8a58d6029bfd588d