MOTECH-1618: Allow setting custom max fetch-depth through the @Entity annotation
The max fetch that will be used for a given entity can now be passed
using the @Entity annotation. This should prevent the DataBrowser from
crashing on non-available fields when the toString() or other methods
depend on deeply nested relationships.

I also fixed some code in the EntityService and cleaned up useless
@Transactional annotations from the InstanceService.

I will document this in another commit, if this is a satisfactory fix.

Change-Id: If8b38cd8bd8568ebd54713c76830d397fc8833d7