MOTECH-1618: Allow setting custom max fetch-depth through the @Entity annotation
The max fetch that will be used for a given entity can now be passed
using the @Entity annotation. This should prevent the DataBrowser from
crashing on non-available fields when the toString() or other methods
depend on deeply nested relationships.

I also fixed some code in the EntityService and cleaned up useless
@Transactional annotations from the InstanceService.

I will document this in another commit, if this is a satisfactory fix.

Change-Id: If8b38cd8bd8568ebd54713c76830d397fc8833d7

Conflicts:
	platform/mds/mds-web/src/main/java/org/motechproject/mds/web/service/impl/InstanceServiceImpl.java
	platform/mds/mds/src/main/java/org/motechproject/mds/osgi/MdsBundleWatcher.java
	platform/mds/mds/src/main/java/org/motechproject/mds/service/EntityService.java
	platform/mds/mds/src/main/java/org/motechproject/mds/service/impl/EntityServiceImpl.java
	platform/mds/mds/src/main/java/org/motechproject/mds/util/Constants.java
	platform/mds/mds/src/test/java/org/motechproject/mds/annotations/internal/EntityProcessorTest.java
	platform/mds/mds/src/test/java/org/motechproject/mds/it/service/EntityServiceContextIT.java