MOTECH-1615: Responsive boostrap startup
This commit changes the bootstrap progress screen
so that it shows a list of modules that are
currently starting. It also displays module
startup errors. Behind the scenes a Blueprint
event handler is used to keep track of started
or failed contexts (OSGi statuses are also tracked).
A service that can be queried for the status
is now exposed by the osgi-platform bundle.
This service in PAX ITs for checking the startup
status, which allows us to drop alot of boilerplate
around the framework startup in tests.

Change-Id: I6c31c6156548578b496ac518a6a4206c73a22c5d