MOTECH-1688 Made MDS respect annotations in M:N relationships
There were several problems with the annotations in the M:N
relationships. The names of the tables would get overwritten
by the generated metadata, since it has got priority over
the values in the annotation. For now, the metadata will
only be generated if it is not already present in the
annotation. Moreover, the mappedBy parameter of
@Persistent annotation is now propagated to the
generated history and trash classes to avoid
problems with relationships there. Finally, the
relationship collections are now determined based on
the collection from main entity, instead of assuming
it's always List.

Change-Id: If4ff7a845245655069ded8e4c486c9a6400a279b