MOTECH-1729: Allow access to mds data browser without MDS permission
This allows to view embedded UI. The verification of permissions
is done on the InstanceService level on a per entity basis. The permission
remains for the purpose of having the root data browser element on the UI
being only accessible to admins. Moreover a dependnecy on a direct call
to web-security has been removed, this call now goes through MDS -
users without access will now receive any suggestions regarding the owner.

Change-Id: I75cd19298fc3ff85276316f9a2ebd5e2e09c242f