MOTECH-1779: Added a permission for UI access
A new permission and a role were added that allow
UI access. Without this permission the user will see a 403
page. This should allow to define users
that should not see the UI at all.

NOTE: this only blocks few of the UI endpoints, making
sure the controllers are secure on the backend is left
to the implementer.

    Change-Id: I3264350d71635880097be886fd23b7b5740d7067

Change-Id: I5ff1fe8ee06ced82336afd5097bb3f1f7579038b