Added @Transactional around some of the InstanceService methods
From now, when the instance is updated, the object retrieval and
modification are taking place within the same transaction. One of the
effects of this is that the object that get passed to the update method
is the exact the same object that was modified. It allows, e.g. to store
transient field with the instance during modification and then access it
in the lifecycle listener.

Additionally, replaced getDeclaredField with FieldUtils.getField in one
more place.

Change-Id: Ia40e746b76edc7527c739ba91aa3bbb6c62edda6