MOTECH-1814: Added methods for retrieval of related fields
Added methods that allow retrieval of related fields in
InstanceService and InstanceController - paging and ordering
can be applied. Currently this is done in memory - improvements
for doing this at the database level can be added later.
When this gets merged it will unblock the other tickets
in the epic for improved related field display.

Change-Id: Ifa4dc5936adde5ff51de28b2746d7e76924239fd