Removed checks for Tomcat processes
I am removing this for now, since with multiple builds running
this can incorectly stop the server or read it as running.
If we have leftover Tomcat processes, we will get to know about it
sooner or later on the CI. In the future we might consider
comming up with a more sophisticated check.