MOTECH-1940: Default UI tab for modules is determined basing on permissions.
- Added tabAccessMap to ModuleRegistrationData. This map contains permissions as keys and
lists of available UI tabs for permission as values. Modules with more than one role can
use it to register which role has access to which tabs.
- ModuleRegistrationData now determines default tab basing on delivered properties. If defaultURL
is specified it will be used as URL of default tab (so all modules with only one role and specified
defaultURL will still works). If not then tabAccessMap should be specified. ModuleRegistrationData
will take first url tahat is accessible for current user permissions and use it as default tab URL.
- Added AvailabilityController (with test) to server-budnle. Modules that have used their own
AvailabilityController now uses this one.