MOTECH-2035: Fixes bug after schema regeneration in MDS
This commit fixes bug, when NucleusException was thrown after schema regeneration
in MDS.

In MDS we use MDS JDO annotation reader, which extends the regular
JDOAnnotationReader from datanucleus. This class was introduced because
regular reader would not read field annotations for metadata if there was
no class level JDO annotations. In this class we fake the PersistenceCapable
annotation.

It works as expected, however sometimes after schema regeneration
datanucleus try to read annotations from class before populate FileMetadata.
In this case we have wrong metadata. We always want to read class annotations
after FileMetada is populated.

After this changes, the bug should be fixed.