MOTECH-2035: Fixes bug after schema regeneration in MDS
This commit fixes bug, when NucleusException was thrown after schema regeneration
in MDS.

In MDS we use MDS JDO annotation reader, which extends the regular
JDOAnnotationReader from datanucleus. This class was introduced because
regular reader would not read field annotations for metadata if there was
no class level JDO annotations. In this class we fake the PersistenceCapable
annotation.

It works as expected, however sometimes after schema regeneration
datanucleus read wrong Metadata for classes. This is connected with this reader.
When we are loading entities using MDSClassLoader our reader implementation is
needed to see field annotations, however after package.jdo file is ready,
there is no need to fake PersistenceCapable annotation. It's why I added the
condition with BundleClassLoader.

After this changes, the bug should be fixed.