MOTECH-2111: Changes color in the error screen to red
There are two exactly the same styles in index.css and bootstrap-page.css.
It means that after building we've got duplicate style in motech.css.
Because of that in the error page there is a problem with overridding this style.
It looks like a background color of the first on is overridden, however the
second duplicated is not. It's why we can see blue color instead of red.

This commit fixes this bug.