MOTECH-2174 Made common DatePicker create child scope, not isolated scope
The common DatePicker was creating an isolated scope and causing the Scheduler module to crash after the Tasks module was loaded. Neither of these modules have their own date-picker module, so the reason for why the Tasks module caused this problem is still unknown.

Additionally, I changed the DatePicker directive to create a child scope because it does use scope variables, and I didn't want to polute potential parent scopes.