MOTECH-2020 Refactored task field bubbles (#357)
* MOTECH-2020 Refactored task field bubbles

Changed tasks field buttons to follow an angularjs design pattern and
less heavily manipulate the DOM. Large HTML files have been split up so
their function is more obvious and readable.

* MOTECH-2020 Changed field format popover to modal

Replaced field format popover with a modal becuase the amount of code
needed to properly display the popover modal was insane. Using a modal
is also more consistent to the MOTECH UI.

* MOTECH-2020 Refactored boolean field bubble

Updated boolean field bubble to use field directive, like other field bubbles. Allowed to get rid of markup specific code in utils and controller functions.

* re-added date-picker code

* MOTECH-2020 Removed draggable and droppable directives.

Removed specific draggable and droppable directives because event passing was causing problems in map interface. Put behavior of draggable and droppable into field and contenteditable directives, respectively.

made single droppbale work in map by removing draggable and droppable directives

* MOTECH-2020 Removed contenteditable markup

There was a bunch of markup inside contenteditable directive implementations which got deleted immedately, so I removed the markup.

* MOTECH-2020 Stopped placeholder text from adding to value when field bubble dropped.

When a field bubble is dropped in the Tasks UI, the placeholder text was getting saved as part of the value.
This was stopped from happening by checking if the field value is unset, and if so, sets the content to the field value.

* MOTECH-2020 Updated contenteditable droppable function to properly check viewValue

There was an error where dragging a dropping a field element would replace any text perviously in the field. This was fixed by checking the correctl varriable name.